feat: Hacked in Buttplu.io support
This uses a .net3.5 "port" of an older version of the c#
buttplug client to send LinearCmd messages based on
the relative distances calculated from the bone positions.

Use Ctrl+i to connect to Intiface

Warning: if the update frequency is too high, the BLE
stack will blow up. There's also some other bugs in there,
possibly related to the dynamic ranges, which casue the
buttplug connection to be dropped.